+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 <!-- There is always Unreleased section on the top. Subsections (Add, Changed, Fix, Removed) should be Add as needed. -->
 ## Unreleased
+- Add `Instance.k8sLocalServiceUrl` function
 
 ## 9.0.0 - 2023-09-11
 - [**BC**] Use `Alma` namespace